chore: consolidate listing drive files with pagination to a common function
