+++ packages/pieces/community/google-drive/src/lib/action/drive-list-files.ts
@@ -1,6 +1,6 @@
 import { googleDriveAuth } from '../auth';
 import { Property, createAction } from '@activepieces/pieces-framework';
-import { getFilesByLevel } from '../common/list-files-recursive';
+import { getFilesByLevel } from '../common/list-drive-files';
 import { downloadFileFromDrive } from '../common/get-file-content';
 import { driveListFilesOutputSchema } from '../output-schemas';
